:bug: Deal with the homepage issue
Going back from solo endpoint to home back to solo would resume the precedent game instead of starting a new one. So If the game was finished (result screen appearing) it was possible to start a new game only by reloading. Which is super inconvinient. The fix is lazy and ugly imo, need improvment.